split runtime control and checkpoint tests
